refactor(gaussdb-pool): Rename `pg-pool` package to `gaussdb-pool`
