+++ packages/gaussdb-pool/esm/index.mjs
@@ -1,4 +1,4 @@
-// ESM wrapper for pg-pool
+// ESM wrapper for gaussdb-pool
 import Pool from '../index.js'
 
 // Export as default only to match CJS module
