+++ packages/gaussdb-pool/test/bring-your-own-promise.js
