+++ packages/gaussdb-pool/test/connection-strings.js
