+++ packages/gaussdb-pool/test/idle-timeout-exit.js
