+++ packages/gaussdb/lib/index.js
@@ -5,7 +5,7 @@ const defaults = require('./defaults')
 const Connection = require('./connection')
 const Result = require('./result')
 const utils = require('./utils')
-const Pool = require('pg-pool')
+const Pool = require('gaussdb-pool')
 const TypeOverrides = require('./type-overrides')
 const { DatabaseError } = require('pg-protocol')
 const { escapeIdentifier, escapeLiteral } = require('./utils')
