+++ packages/pg-esm-test/package.json
@@ -17,7 +17,7 @@
     "gaussdb-cloudflare": "^1.2.5",
     "gaussdb-cursor": "^2.15.0",
     "pg-native": "^3.5.0",
-    "pg-pool": "^3.10.0",
+    "gaussdb-pool": "^3.10.0",
     "pg-protocol": "^1.10.0",
     "pg-query-stream": "^4.10.0"
   },
