+++ packages/pg-esm-test/pg-pool.test.js
@@ -1,8 +1,8 @@
 import assert from 'node:assert'
 import { describe, it } from 'node:test'
-import Pool from 'pg-pool'
+import Pool from 'gaussdb-pool'
 
-describe('pg-pool', () => {
+describe('gaussdb-pool', () => {
   it('should export Pool constructor', () => {
     assert.ok(new Pool())
   })
